requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
requests
python-multipart

streamlit

langchain
langchain-community
langchain-openai
langchain-tavily
langgraph

langchain-redis
redis

sqlalchemy
psycopg2-binary

faiss-cpu
numpy
rank-bm25
sentence-transformers

pypdf
python-docx
